requirements.txt
-e .
coverage
mockredispy
django-redis-cache==2.1.1
celery<5
